Fix a very weird bug
`colmax` seems to return canonical row indices, and not indices that are
sorted wrt. the stack.  This is probably due to having set `row_perm`
somewhere to something wrong.  When visualizing the birth-death pairs
from the points where the initial reduction was done, things are okay.
However, looking at other points, it's all messed up.